Removed unused components. Refactored code
